Restored to '35dee3077f32392e2ac484d1cd771c7459170e41'

+++ shared/schema.ts
@@ -20,7 +20,7 @@ export const insertDonationSchema = createInsertSchema(donations).omit({
   status: true,
   createdAt: true,
 }).extend({
-  amount: z.number().int().min(1),
+  amount: z.number().min(1),
   name: z.string().min(3),
   email: z.string().email(),
   cpf: z.string().min(11).max(14),
